Consolidate duplicate test data
