+++ HousingRepairsOnlineApi.Tests/HelpersTests/EnvironmentVariableRepairTypeSorConfigurationProviderTests.cs
@@ -8,7 +8,7 @@ namespace HousingRepairsOnlineApi.Tests.HelpersTests;
 public class EnvironmentVariableRepairTypeSorConfigurationProviderTests
 {
     [Theory]
-    [MemberData(nameof(InvalidArgumentTestData))]
+    [MemberData(nameof(Helpers.RepairTypeTestData.InvalidRepairTypeArgumentTestData), MemberType = typeof(Helpers.RepairTypeTestData))]
 #pragma warning disable xUnit1026
     public void GivenAnInvalidRepairType_WhenConstructing_ThenExceptionIsThrown<T>(T exception, string repairTypeParameter) where T : Exception
 #pragma warning restore xUnit1026
@@ -22,18 +22,9 @@ public class EnvironmentVariableRepairTypeSorConfigurationProviderTests
         act.Should().ThrowExactly<T>();
     }
 
-    public static TheoryData<Exception, string> InvalidArgumentTestData() =>
-        new()
-        {
-            { new ArgumentNullException(), null },
-            { new ArgumentException(), "" },
-            { new ArgumentException(), " " },
-            { new ArgumentException(), "non-repair-type-value" },
-        };
-
     [Theory]
 
-    [MemberData(nameof(ValidRepairTypeArgumentTestData))]
+    [MemberData(nameof(Helpers.RepairTypeTestData.ValidRepairTypeArgumentTestData), MemberType = typeof(Helpers.RepairTypeTestData))]
     public void GivenValidRepairTypeParameter_WhenConstructing_ThenExceptionIsNotThrown(string repairTypeParameter)
     {
         // Arrange
@@ -43,15 +34,5 @@ public class EnvironmentVariableRepairTypeSorConfigurationProviderTests
 
         // Assert
         act.Should().NotThrow<ArgumentException>();
-    }
-
-    public static TheoryData<string> ValidRepairTypeArgumentTestData()
-    {
-        var result = new TheoryData<string>();
-        foreach (var repairType in RepairType.All)
-        {
-            result.Add(repairType);
-        }
-        return result;
     }
 }
