+++ HousingRepairsOnlineApi.Tests/UseCasesTests/RetrieveAvailableAppointmentsUseCaseTests.cs
@@ -94,8 +94,7 @@ namespace HousingRepairsOnlineApi.Tests.UseCasesTests
         }
 
         [Theory]
-        [MemberData(nameof(InvalidArgumentTestData))]
-        [MemberData(nameof(InvalidRepairTypeArgument))]
+        [MemberData(nameof(Helpers.RepairTypeTestData.InvalidRepairTypeArgumentTestData), MemberType = typeof(Helpers.RepairTypeTestData))]
 #pragma warning disable xUnit1026
         public async void GivenAnInvalidRepairType_WhenExecute_ThenExceptionIsThrown<T>(T exception, string repairTypeParameter) where T : Exception
 #pragma warning restore xUnit1026
@@ -116,10 +115,8 @@ namespace HousingRepairsOnlineApi.Tests.UseCasesTests
             yield return new object[] { new ArgumentException(), " " };
         }
 
-        public static TheoryData<Exception, string> InvalidRepairTypeArgument() => new() { { new ArgumentException(), "non-repair-type-value" } };
-
         [Theory]
-        [MemberData(nameof(ValidRepairTypeArgumentTestData))]
+        [MemberData(nameof(Helpers.RepairTypeTestData.ValidRepairTypeArgumentTestData), MemberType = typeof(Helpers.RepairTypeTestData))]
         public void GivenValidRepairTypeParameter_WhenResolving_ThenExceptionIsNotThrown(string repairTypeParameter)
         {
             // Arrange
@@ -129,16 +126,6 @@ namespace HousingRepairsOnlineApi.Tests.UseCasesTests
 
             // Assert
             act.Should().NotThrow<ArgumentException>();
-        }
-
-        public static TheoryData<string> ValidRepairTypeArgumentTestData()
-        {
-            var result = new TheoryData<string>();
-            foreach (var repairType in RepairType.All)
-            {
-                result.Add(repairType);
-            }
-            return result;
         }
 
         [Fact]
